Moving towards SVN integration.
I like my version number to be unique and easy to update, so that when I'm
cranking on things mid-semester, I don't get confused when communicating with
students.

I'm looking to get SVN commit info in the document itself, so we can say "You
should have revision XYZ," and not have students be *too* confused about what
is going on.