Versioning the filename of the build output.
When I post the PDF, I want it to carry the version info in the filename. 

I've modified the Makefile to do this, but note... if you are building from an
uncommitted version, you'll get something like this:

android4cs2-rev134:135M.pdf

That's not good, but... I'll keep it for now. The reason is this: we should
probably not hand out a version of the document to students that is not
committed. This will be a visual indicator that you're between commits.